requirements.txt
fastapi
uvicorn[standard]
mss
Pillow
python-multipart
